+++ src/driver.cpp
@@ -19,29 +19,29 @@ void Driver::assignment(uint64_t key, uint64_t value)
 		return;
 	}
 
-	int nodes = bst.assignment(key, value);
+	uint32_t nodes = bst.assignment(key, value);
 	out << "wezlow: " << nodes << '\n';
 }
 
 void Driver::sum(uint64_t time, uint64_t left, uint64_t right)
 {
-	if (time > (uint64_t)bst.now() || left > right || right > maxKey || !stmtCheck()) {
+	if (time > bst.now() || left > right || right > maxKey || !stmtCheck()) {
 		error();
 		return;
 	}
 
-	int sum = bst.sum(time, left, right);
+	uint64_t sum = bst.sum(time, left, right);
 	out << "suma(" << time << ',' << left << ".." << right << ")=" << sum << '\n';
 }
 
 void Driver::clear(uint64_t time)
 {
-	if (time > (uint64_t)bst.now() || !stmtCheck()) {
+	if (time > bst.now() || !stmtCheck()) {
 		error();
 		return;
 	}
 
-	int nodes = bst.clear(time);
+	uint32_t nodes = bst.clear(time);
 	out << "wezlow: " << nodes << '\n';
 }
